Use redux forms for skip logic editor
