+++ src/components/Query/index.js
@@ -2,3 +2,4 @@ export { default as Filter } from './Filter';
 export { default as Query } from './Query';
 export { default as withStoreConnector } from './withStoreConnector';
 export { default as withFieldConnector } from './withFieldConnector';
+export { default as ruleValidator } from './ruleValidator';
